@@ -5,7 +5,7 @@ import (
 	"crypto/tls"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net"
 	"net/http"
 	"strconv"
@@ -64,6 +64,10 @@ func createUpstreamClient(cfg config.Upstream) upstreamClient {
 		MinVersion: tls.VersionTLS12,
 	}
 
+	if cfg.CommonName != "" {
+		tlsConfig.ServerName = cfg.CommonName
+	}
+
 	switch cfg.Net {
 	case config.NetProtocolHttps:
 		return &httpUpstreamClient{
@@ -109,7 +113,7 @@ func createUpstreamClient(cfg config.Upstream) upstreamClient {
 }
 
 func (r *httpUpstreamClient) fmtURL(ip net.IP, port uint16, path string) string {
-	return fmt.Sprintf("https://%s:%d%s", ip.String(), port, path)
+	return fmt.Sprintf("https://%s%s", net.JoinHostPort(ip.String(), strconv.Itoa(int(port))), path)
 }
 
 func (r *httpUpstreamClient) callExternal(msg *dns.Msg,
@@ -152,7 +156,7 @@ func (r *httpUpstreamClient) callExternal(msg *dns.Msg,
 			dnsContentType, contentType)
 	}
 
-	body, err := ioutil.ReadAll(httpResponse.Body)
+	body, err := io.ReadAll(httpResponse.Body)
 	if err != nil {
 		return nil, 0, fmt.Errorf("can't read response body:  %w", err)
 	}
@@ -167,7 +171,7 @@ func (r *httpUpstreamClient) callExternal(msg *dns.Msg,
 	return &response, time.Since(start), nil
 }
 
-func (r *dnsUpstreamClient) fmtURL(ip net.IP, port uint16, _path string) string {
+func (r *dnsUpstreamClient) fmtURL(ip net.IP, port uint16, _ string) string {
 	return net.JoinHostPort(ip.String(), strconv.Itoa(int(port)))
 }
 
