@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"encoding/base64"
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net"
 	"net/http"
 	"strings"
@@ -61,8 +61,15 @@ var _ = BeforeSuite(func() {
 	DeferCleanup(fritzboxMockUpstream.Close)
 
 	clientMockUpstream := resolver.NewMockUDPUpstreamServer().WithAnswerFn(func(request *dns.Msg) (response *dns.Msg) {
+		var clientName string
+		client := mockClientName.Load()
+
+		if client != nil {
+			clientName = mockClientName.Load().(string)
+		}
+
 		response, err := util.NewMsgWithAnswer(
-			util.ExtractDomain(request.Question[0]), 3600, dns.Type(dns.TypePTR), mockClientName.Load().(string),
+			util.ExtractDomain(request.Question[0]), 3600, dns.Type(dns.TypePTR), clientName,
 		)
 
 		Expect(err).Should(Succeed())
@@ -74,6 +81,28 @@ var _ = BeforeSuite(func() {
 	upstreamClient = clientMockUpstream.Start()
 	upstreamFritzbox = fritzboxMockUpstream.Start()
 	upstreamGoogle = googleMockUpstream.Start()
+
+	tmpDir := NewTmpFolder("server")
+	Expect(tmpDir.Error).Should(Succeed())
+	DeferCleanup(tmpDir.Clean)
+
+	certPem := writeCertPem(tmpDir)
+	Expect(certPem.Error).Should(Succeed())
+
+	keyPem := writeKeyPem(tmpDir)
+	Expect(keyPem.Error).Should(Succeed())
+
+	doubleclickFile := tmpDir.CreateStringFile("doubleclick.net.txt", "doubleclick.net", "doubleclick.net.cn")
+	Expect(doubleclickFile.Error).Should(Succeed())
+
+	bildFile := tmpDir.CreateStringFile("www.bild.de.txt", "www.bild.de")
+	Expect(bildFile.Error).Should(Succeed())
+
+	heiseFile := tmpDir.CreateStringFile("heise.de.txt", "heise.de")
+	Expect(heiseFile.Error).Should(Succeed())
+
+	youtubeFile := tmpDir.CreateStringFile("youtube.com.txt", "youtube.com")
+	Expect(youtubeFile.Error).Should(Succeed())
 
 	// create server
 	sut, err = NewServer(&config.Config{
@@ -97,13 +126,13 @@ var _ = BeforeSuite(func() {
 		Blocking: config.BlockingConfig{
 			BlackLists: map[string][]string{
 				"ads": {
-					"../testdata/doubleclick.net.txt",
-					"../testdata/www.bild.de.txt",
-					"../testdata/heise.de.txt"},
-				"youtube": {"../testdata/youtube.com.txt"}},
+					doubleclickFile.Path,
+					bildFile.Path,
+					heiseFile.Path},
+				"youtube": {youtubeFile.Path}},
 			WhiteLists: map[string][]string{
-				"ads":       {"../testdata/heise.de.txt"},
-				"whitelist": {"../testdata/heise.de.txt"},
+				"ads":       {heiseFile.Path},
+				"whitelist": {heiseFile.Path},
 			},
 			ClientGroupsBlock: map[string][]string{
 				"default":         {"ads"},
@@ -123,8 +152,8 @@ var _ = BeforeSuite(func() {
 
 		DNSPorts:   config.ListenConfig{"55555"},
 		TLSPorts:   config.ListenConfig{"8853"},
-		CertFile:   "../testdata/cert.pem",
-		KeyFile:    "../testdata/key.pem",
+		CertFile:   certPem.Path,
+		KeyFile:    keyPem.Path,
 		HTTPPorts:  config.ListenConfig{"4000"},
 		HTTPSPorts: config.ListenConfig{"4443"},
 		Prometheus: config.PrometheusConfig{
@@ -388,7 +417,7 @@ var _ = Describe("Running DNS server", func() {
 					Expect(resp).Should(HaveHTTPStatus(http.StatusOK))
 					Expect(resp).Should(HaveHTTPHeaderWithValue("Content-type", "application/dns-message"))
 
-					rawMsg, err := ioutil.ReadAll(resp.Body)
+					rawMsg, err := io.ReadAll(resp.Body)
 					Expect(err).Should(Succeed())
 
 					msg := new(dns.Msg)
@@ -451,7 +480,7 @@ var _ = Describe("Running DNS server", func() {
 					defer resp.Body.Close()
 					Expect(resp).Should(HaveHTTPStatus(http.StatusOK))
 					Expect(resp).Should(HaveHTTPHeaderWithValue("Content-type", "application/dns-message"))
-					rawMsg, err := ioutil.ReadAll(resp.Body)
+					rawMsg, err := io.ReadAll(resp.Body)
 					Expect(err).Should(Succeed())
 
 					msg = new(dns.Msg)
@@ -471,7 +500,7 @@ var _ = Describe("Running DNS server", func() {
 					defer resp.Body.Close()
 					Expect(resp).Should(HaveHTTPStatus(http.StatusOK))
 					Expect(resp).Should(HaveHTTPHeaderWithValue("Content-type", "application/dns-message"))
-					rawMsg, err := ioutil.ReadAll(resp.Body)
+					rawMsg, err := io.ReadAll(resp.Body)
 					Expect(err).Should(Succeed())
 
 					msg = new(dns.Msg)
@@ -528,7 +557,7 @@ var _ = Describe("Running DNS server", func() {
 			Expect(cErr).Should(Succeed())
 
 			cfg.Upstream.ExternalResolvers = map[string][]config.Upstream{
-				"default": {config.Upstream{Net: config.NetProtocolTcpUdp, Host: "4.4.4.4", Port: 53}}}
+				"default": {config.Upstream{Net: config.NetProtocolTcpUdp, Host: "1.1.1.1", Port: 53}}}
 
 			cfg.Redis.Address = "test-fail"
 		})
@@ -658,7 +687,7 @@ var _ = Describe("Running DNS server", func() {
 			Expect(cErr).Should(Succeed())
 
 			cfg.Upstream.ExternalResolvers = map[string][]config.Upstream{
-				"default": {config.Upstream{Net: config.NetProtocolTcpUdp, Host: "4.4.4.4", Port: 53}}}
+				"default": {config.Upstream{Net: config.NetProtocolTcpUdp, Host: "1.1.1.1", Port: 53}}}
 		})
 
 		It("should create self-signed certificate if key/cert files are not provided", func() {
@@ -706,4 +735,42 @@ func requestServer(request *dns.Msg) *dns.Msg {
 	Log().Fatal("could not read from connection", err)
 
 	return nil
+}
+
+func writeCertPem(tmpDir *TmpFolder) *TmpFile {
+	return tmpDir.CreateStringFile("cert.pem",
+		"-----BEGIN CERTIFICATE-----",
+		"MIICMzCCAZygAwIBAgIRAJCCrDTGEtZfRpxDY1KAoswwDQYJKoZIhvcNAQELBQAw",
+		"EjEQMA4GA1UEChMHQWNtZSBDbzAgFw03MDAxMDEwMDAwMDBaGA8yMDg0MDEyOTE2",
+		"MDAwMFowEjEQMA4GA1UEChMHQWNtZSBDbzCBnzANBgkqhkiG9w0BAQEFAAOBjQAw",
+		"gYkCgYEA4mEaF5yWYYrTfMgRXdBpgGnqsHIADQWlw7BIJWD/gNp+fgp4TUZ/7ggV",
+		"rrvRORvRFjw14avd9L9EFP7XLi8ViU3uoE1UWI32MlrKqLbGNCXyUIApIoqlbRg6",
+		"iErxIk5+ChzFuysQOx01S2yv/ML6dx7NOGHs1S38MUzRZtcXBH8CAwEAAaOBhjCB",
+		"gzAOBgNVHQ8BAf8EBAMCAqQwEwYDVR0lBAwwCgYIKwYBBQUHAwEwDwYDVR0TAQH/",
+		"BAUwAwEB/zAdBgNVHQ4EFgQUslNI6tYIv909RttHaZVMS/u/VYYwLAYDVR0RBCUw",
+		"I4IJbG9jYWxob3N0hwR/AAABhxAAAAAAAAAAAAAAAAAAAAABMA0GCSqGSIb3DQEB",
+		"CwUAA4GBAJ2gRpQHr5Qj7dt26bYVMdN4JGXTsvjbVrJfKI0VfPGJ+SUY/uTVBUeX",
+		"+Cwv4DFEPBlNx/lzuUkwmRaExC4/w81LWwxe5KltYsjyJuYowiUbLZ6tzLaQ9Bcx",
+		"jxClAVvgj90TGYOwsv6ESOX7GWteN1FlD3+jk7vefjFagaKKFYR9",
+		"-----END CERTIFICATE-----")
+}
+
+func writeKeyPem(tmpDir *TmpFolder) *TmpFile {
+	return tmpDir.CreateStringFile("key.pem",
+		"-----BEGIN PRIVATE KEY-----",
+		"MIICeAIBADANBgkqhkiG9w0BAQEFAASCAmIwggJeAgEAAoGBAOJhGheclmGK03zI",
+		"EV3QaYBp6rByAA0FpcOwSCVg/4Dafn4KeE1Gf+4IFa670Tkb0RY8NeGr3fS/RBT+",
+		"1y4vFYlN7qBNVFiN9jJayqi2xjQl8lCAKSKKpW0YOohK8SJOfgocxbsrEDsdNUts",
+		"r/zC+ncezThh7NUt/DFM0WbXFwR/AgMBAAECgYEA1exixstPhI+2+OTrHFc1S4dL",
+		"oz+ncqbSlZEBLGl0KWTQQfVM5+FmRR7Yto1/0lLKDBQL6t0J2x3fjWOhHmCaHKZA",
+		"VAvZ8+OKxwofih3hlO0tGCB8szUJygp2FAmd0rOUqvPQ+PTohZEUXyDaB8MOIbX+",
+		"qoo7g19+VlbyKqmM8HkCQQDs4GQJwEn7GXKllSMyOfiYnjQM2pwsqO0GivXkH+p3",
+		"+h5KDp4g3O4EbmbrvZyZB2euVsBjW3pFMu+xPXuOXf91AkEA9KfC7LGLD2OtLmrM",
+		"iCZAqHlame+uEEDduDmqjTPnNKUWVeRtYKMF5Hltbeo1jMXMSbVZ+fRWKfQ+HAhQ",
+		"xjFJowJAV6U7PqRoe0FSO1QwXrA2fHnk9nCY4qlqckZObyckAVqJhIteFPjKFNeo",
+		"u0dAPxsPUOGGc/zwA9Sx/ZmrMuUy1QJBALl7bqawO/Ng6G0mfwZBqgeQaYYHVnnw",
+		"E6iV353J2eHpvzNDSUFYlyEOhk4soIindSf0m9CK08Be8a+jBkocF+0CQQC+Hi7L",
+		"kZV1slpW82BxYIhs9Gb0OQgK8SsI4aQPTFGUarQXXAm4eRqBO0kaG+jGX6TtW353",
+		"EHK784GIxwVXKej/",
+		"-----END PRIVATE KEY-----")
 }
