@@ -9,61 +9,102 @@ import (
 	"github.com/miekg/dns"
 	"github.com/sirupsen/logrus"
 	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 
 	. "github.com/onsi/gomega"
 
 	. "github.com/onsi/ginkgo/v2"
 )
 
+var err error
+
 var _ = Describe("DatabaseWriter", func() {
 
-	Describe("Database query log", func() {
-		When("New log entry was created", func() {
-			It("should be persisted in the database", func() {
-				sqlite := sqlite.Open("file::memory:")
-				writer, err := newDatabaseWriter(sqlite, 7, time.Millisecond)
-				Expect(err).Should(Succeed())
-				request := &model.Request{
-					Req: util.NewMsgWithQuestion("google.de.", dns.Type(dns.TypeA)),
-					Log: logrus.NewEntry(log.Log()),
-				}
-				res, err := util.NewMsgWithAnswer("example.com", 123, dns.Type(dns.TypeA), "123.124.122.122")
+	Describe("Database query log to sqlite", func() {
+		var (
+			sqliteDB gorm.Dialector
+			writer   *DatabaseWriter
+			request  *model.Request
+		)
 
+		BeforeEach(func() {
+			sqliteDB = sqlite.Open("file::memory:")
+
+			request = &model.Request{
+				Req: util.NewMsgWithQuestion("google.de.", dns.Type(dns.TypeA)),
+				Log: logrus.NewEntry(log.Log()),
+			}
+		})
+
+		When("New log entry was created", func() {
+			BeforeEach(func() {
+				writer, err = newDatabaseWriter(sqliteDB, 7, time.Millisecond)
 				Expect(err).Should(Succeed())
+			})
+
+			It("should be persisted in the database", func() {
+				res, err := util.NewMsgWithAnswer("example.com", 123, dns.Type(dns.TypeA), "123.124.122.122")
+				Expect(err).Should(Succeed())
+
 				response := &model.Response{
 					Res:    res,
 					Reason: "Resolved",
 					RType:  model.ResponseTypeRESOLVED,
 				}
+
+				// one entry with now as timestamp
 				writer.Write(&LogEntry{
 					Request:    request,
 					Response:   response,
 					Start:      time.Now(),
 					DurationMs: 20,
 				})
+
+				// one entry before 2 days
+				writer.Write(&LogEntry{
+					Request:    request,
+					Response:   response,
+					Start:      time.Now().AddDate(0, 0, -2),
+					DurationMs: 20,
+				})
+
+				// force write
+				writer.doDBWrite()
+
+				// 2 entries in the database
+				Eventually(func() int64 {
+					var res int64
+					result := writer.db.Find(&logEntry{})
+
+					result.Count(&res)
+
+					return res
+				}, "5s").Should(BeNumerically("==", 2))
+
+				// do cleanup now
+				writer.CleanUp()
+
+				// now only 1 entry in the database
 				Eventually(func() (res int64) {
 					result := writer.db.Find(&logEntry{})
 
 					result.Count(&res)
 
 					return res
-				}, "1s").Should(BeNumerically("==", 1))
+				}, "5s").Should(BeNumerically("==", 2))
 			})
 		})
 
 		When("There are log entries with timestamp exceeding the retention period", func() {
+			BeforeEach(func() {
+				writer, err = newDatabaseWriter(sqliteDB, 1, time.Millisecond)
+				Expect(err).Should(Succeed())
+			})
+
 			It("these old entries should be deleted", func() {
-				sqlite := sqlite.Open("file::memory:")
-				writer, err := newDatabaseWriter(sqlite, 1, time.Millisecond)
-				Expect(err).Should(Succeed())
-
-				request := &model.Request{
-					Req: util.NewMsgWithQuestion("google.de.", dns.Type(dns.TypeA)),
-					Log: logrus.NewEntry(log.Log()),
-				}
 				res, err := util.NewMsgWithAnswer("example.com", 123, dns.Type(dns.TypeA), "123.124.122.122")
-
 				Expect(err).Should(Succeed())
+
 				response := &model.Response{
 					Res:    res,
 					Reason: "Resolved",
@@ -86,6 +127,9 @@ var _ = Describe("DatabaseWriter", func() {
 					DurationMs: 20,
 				})
 
+				// force write
+				writer.doDBWrite()
+
 				// 2 entries in the database
 				Eventually(func() int64 {
 					var res int64
@@ -94,7 +138,7 @@ var _ = Describe("DatabaseWriter", func() {
 					result.Count(&res)
 
 					return res
-				}, "1s").Should(BeNumerically("==", 2))
+				}, "5s").Should(BeNumerically("==", 2))
 
 				// do cleanup now
 				writer.CleanUp()
@@ -106,10 +150,12 @@ var _ = Describe("DatabaseWriter", func() {
 					result.Count(&res)
 
 					return res
-				}, "1s").Should(BeNumerically("==", 1))
+				}, "5s").Should(BeNumerically("==", 1))
 			})
 		})
+	})
 
+	Describe("Database query log fails", func() {
 		When("mysql connection parameters wrong", func() {
 			It("should be log with fatal", func() {
 				_, err := NewDatabaseWriter("mysql", "wrong param", 7, 1)
