@@ -26,7 +26,8 @@ var (
 
 // Bootstrap allows resolving hostnames using the configured bootstrap DNS.
 type Bootstrap struct {
-	log *logrus.Entry
+	log                 *logrus.Entry
+	startVerifyUpstream bool
 
 	resolver    Resolver
 	upstream    Resolver // the upstream that's part of the above resolver
@@ -64,9 +65,10 @@ func NewBootstrap(cfg *config.Config) (b *Bootstrap, err error) {
 	// This also prevents the GC to clean up these two structs, but is not currently an
 	// issue since they stay allocated until the process terminates
 	b = &Bootstrap{
-		log:            log,
-		upstreamIPs:    ips,
-		systemResolver: net.DefaultResolver, // allow replacing it during tests
+		log:                 log,
+		upstreamIPs:         ips,
+		systemResolver:      net.DefaultResolver, // allow replacing it during tests
+		startVerifyUpstream: cfg.StartVerifyUpstream,
 	}
 
 	if upstream.IsDefault() {
@@ -96,26 +98,18 @@ func (b *Bootstrap) UpstreamIPs(r *UpstreamResolver) (*IPSet, error) {
 func (b *Bootstrap) resolveUpstream(r Resolver, host string) ([]net.IP, error) {
 	// Use system resolver if no bootstrap is configured
 	if b.resolver == nil {
-		filteredQTypes := config.GetConfig().Filtering.QueryTypes
-
-		network := "ip"
-		if filteredQTypes.Contains(dns.Type(dns.TypeAAAA)) {
-			network = "ip4"
-		} else if filteredQTypes.Contains(dns.Type(dns.TypeA)) {
-			network = "ip6"
-		}
-
+		cfg := config.GetConfig()
 		ctx := context.Background()
 
-		timeout := config.GetConfig().UpstreamTimeout
+		timeout := cfg.UpstreamTimeout
 		if timeout != 0 {
 			var cancel context.CancelFunc
 
-			ctx, cancel = context.WithTimeout(context.Background(), time.Duration(timeout))
+			ctx, cancel = context.WithTimeout(ctx, time.Duration(timeout))
 			defer cancel()
 		}
 
-		return b.systemResolver.LookupIP(ctx, network, host)
+		return b.systemResolver.LookupIP(ctx, cfg.ConnectIPVersion.Net(), host)
 	}
 
 	if r == b.upstream {
@@ -145,14 +139,16 @@ func (b *Bootstrap) NewHTTPTransport() *http.Transport {
 				return nil, err
 			}
 
-			filteredQTypes := config.GetConfig().Filtering.QueryTypes
+			connectIPVersion := config.GetConfig().ConnectIPVersion
 
 			var qTypes []dns.Type
 
 			switch {
-			case strings.HasSuffix(network, "4") || filteredQTypes.Contains(dns.Type(dns.TypeAAAA)):
+			case connectIPVersion != config.IPVersionDual: // ignore `network` if a specific version is configured
+				qTypes = connectIPVersion.QTypes()
+			case strings.HasSuffix(network, "4"):
 				qTypes = []dns.Type{dns.Type(dns.TypeA)}
-			case strings.HasSuffix(network, "6") || filteredQTypes.Contains(dns.Type(dns.TypeA)):
+			case strings.HasSuffix(network, "6"):
 				qTypes = []dns.Type{dns.Type(dns.TypeAAAA)}
 			default:
 				qTypes = v4v6QTypes
